refactor: page refactor
